requirements.txt
azure-functions
langchain>=0.0.278
langchain-openai>=0.0.2
langchain-community>=0.0.2
openai==0.28.1
tiktoken==0.4.0
python-multipart==0.0.6
numpy>=1.22.0
pinecone-client>=2.2.1